requirements.txt
protobuf==3.20.3
grpcio==1.48.2
grpcio-tools==1.48.2
etcd3==0.12.0
kafka-python==2.0.2